Create module_utils.common.sentinal
