+++ lib/ansible/playbook/taggable.py
@@ -19,9 +19,9 @@ from __future__ import annotations
 
 from ansible.errors import AnsibleError
 from ansible.module_utils.six import string_types
+from ansible.module_utils.common.sentinel import Sentinel
 from ansible.playbook.attribute import FieldAttribute
 from ansible.template import Templar
-from ansible.utils.sentinel import Sentinel
 
 
 def _flatten_tags(tags: list) -> list:
