+++ lib/ansible/plugins/lookup/config.py
@@ -84,10 +84,10 @@ import ansible.plugins.loader as plugin_loader
 
 from ansible import constants as C
 from ansible.errors import AnsibleError, AnsibleLookupError, AnsibleOptionsError
+from ansible.module_utils.common.sentinel import Sentinel
 from ansible.module_utils.common.text.converters import to_native
 from ansible.module_utils.six import string_types
 from ansible.plugins.lookup import LookupBase
-from ansible.utils.sentinel import Sentinel
 
 
 class MissingSetting(AnsibleOptionsError):
